gpdiff.pl: only quote regular expression with \Q and \E
The script gpdiff.pl will replace file path in installcheck output
with a fixed path, then produce the diff result.
On windows the path separator is backslach, so we need to quote
the regex, which is the file path, with \Q and \E. We don't have
to do this with the replacement side.