Allow exchanging partition of a partition table
Before PG12 merge, we can exchange a partition of a partition table
with the following syntax:
ALTER TABLE XXX
ALTER PARTITION XXX
EXCHANGE PARTITION XXX
WITH TABLE XXX;

It's now rejected with message "table XXX is not partitioned". This
is probably misleading. Actually this error messages is intended for
ADD or DROP partition table, so EXCHANGE PARTITION should also be
an exception. Then EXCHANGE PARTITION syntax should work.

But the behavior is still slightly different to 6X. Previously the
"ALTER PARTITION" can be specified even for 1 level partition table,
now it's not allowed any more. "ALTER PARTITION" should only be used
to exchange subpartitions of partition table with 2 or more levels.