Move most external table code to extension
As we are migrating from external table to FDW, code at
src/backend/access/external is no longer needed in kernel.
Because we still need to support the syntax and the extprotocol
interface(for gpcloud), we still leave the command implementation
(exttablecmds.c) and CreateExternalStmt node(parsenodes.h) in
kernel. They will remain until we can finally remove syntax support
for external table.

Other notes: gpdb specific environment is no longer supported
in COPY TO/FROM command. This feature is for execute external
table, it's not documented for COPY command anyway.