Fix various issues.
- Remove role GP_ROLE_RETRIEVE.  This role would be quite confusing for users
that needs to write gp_role related code and do not understand the feature.  We
might use gpconntype instead - that could make the code simpler, but that is
not a standard parameter in various clients.  Finally decide that we still use
bool guc gp_retrieve_conn, and run the retrieve statement in utility mode.  We
may still use the RETRIEVE role internally but for both cases (RETRIEVE and
UTILITY), we still need to worry about the using. We have run a lot for the
utility role so let's use the utility role at first and then see if we wan to
use the new role.

- Fix some bugs related to previous GP_ROLE_RETRIEVE code.  One is if the guc
is specified multiple times, another is it should allow more values for bool
types (e.g. true, TRUE, 1, etc), besides guc_options are not considered. To be
safer, a sanity check code is added to ensure retrieve connection must be
authenticated.

- Various other code tweaking.