Refactor endpoint code (round 2. mainly cdbendpoint.c) and fix issues.
Main fix is: Remove the hash entry if it's ref count is 0. Previously this was
just done during abort - this is wrong. clean_session_token_info() is removed
accordingly.

Token includes gp_session_id + random bytes, the comment said:
"Token is unique for every session id. This is guaranteed by using the session
 id as a part of the token. And same session will have the same token. Thus the
 retriever will know which session to attach when doing authentication."
but the session id is not needed since even there is random bytes collision
it's not a big issue.

Code refactor change please see the diff for details.